+++ src/engine/InputQuery.h
@@ -22,7 +22,7 @@
 #include "Map.h"
 #include "NetworkLevelReasoner.h"
 #include "PiecewiseLinearConstraint.h"
-#include "TranscendentalConstraint.h"
+#include "NonlinearConstraint.h"
 
 class InputQuery
 {
@@ -54,9 +54,9 @@ public:
     const List<PiecewiseLinearConstraint *> &getPiecewiseLinearConstraints() const;
     List<PiecewiseLinearConstraint *> &getPiecewiseLinearConstraints();
   
-    void addTranscendentalConstraint( TranscendentalConstraint *constraint );
-    const List<TranscendentalConstraint *> &getTranscendentalConstraints() const;
-    List<TranscendentalConstraint *> &getTranscendentalConstraints();
+    void addNonlinearConstraint( NonlinearConstraint *constraint );
+    const List<NonlinearConstraint *> &getNonlinearConstraints() const;
+    List<NonlinearConstraint *> &getNonlinearConstraints();
 
     /*
       Methods for handling input and output variables
@@ -149,7 +149,7 @@ private:
     Map<unsigned, double> _lowerBounds;
     Map<unsigned, double> _upperBounds;
     List<PiecewiseLinearConstraint *> _plConstraints;
-    List<TranscendentalConstraint *> _tsConstraints;
+    List<NonlinearConstraint *> _nlConstraints;
 
     Map<unsigned, double> _solution;
 
@@ -179,6 +179,12 @@ private:
     bool constructMaxLayer( NLR::NetworkLevelReasoner *nlr,
                             Map<unsigned, unsigned> &handledVariableToLayer,
                             unsigned newLayerIndex );
+    bool constructBilinearLayer( NLR::NetworkLevelReasoner *nlr,
+                                  Map<unsigned, unsigned> &handledVariableToLayer,
+                                  unsigned newLayerIndex );
+    bool constructSoftmaxLayer( NLR::NetworkLevelReasoner *nlr,
+                                Map<unsigned, unsigned> &handledVariableToLayer,
+                                unsigned newLayerIndex );
 
 public:
     /*