+++ src/engine/MILPEncoder.h
@@ -16,12 +16,14 @@
 #ifndef __MILPEncoder_h__
 #define __MILPEncoder_h__
 
+#include "BilinearConstraint.h"
 #include "DisjunctionConstraint.h"
 #include "GurobiWrapper.h"
 #include "InputQuery.h"
 #include "ITableau.h"
 #include "LinearExpression.h"
 #include "MStringf.h"
+#include "SoftmaxConstraint.h"
 #include "Statistics.h"
 
 #include "Map.h"
@@ -134,6 +136,18 @@ private:
       Encode a Sigmoid constraint
     */
     void encodeSigmoidConstraint( GurobiWrapper &gurobi, SigmoidConstraint *sigmoid );
+
+    /*
+      Encode a Softmax constraint
+    */
+    void encodeSoftmaxConstraint( GurobiWrapper &gurobi, SoftmaxConstraint *softmax );
+
+    /*
+      Encode a Bilinear constraint
+    */
+    void encodeBilinearConstraint( GurobiWrapper &gurobi,
+                                   BilinearConstraint *bilinear,
+                                   bool relax );
 };
 
 #endif // __MILPEncoder_h__