+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Added support for onnx networks with tanh nodes to command-line interface.
 * Added proof producing versions of Sign, Max, Absolute Value and Disjunction constraints
 * Added support for properties provided in VNN-LIB format for ONNX networks via the Python API.
+* Support additional non-linear constraints Softmax and Bilinear
 
 ## Version 1.0.0
 * Initial versioned release